fix: MSVC 2026 の C4242 残存警告を解消（追加箇所）
GitHub Actions のログ出力上限で見落とされていた箇所を、ローカルの
VS2026 ビルドで検出して修正。

- monster-damage.cpp:428: speed_to_energy(target.speed) の引数を
  byte に明示キャスト。
- save/monster-writer.cpp:235: PERCENTAGE 型 transform_hp_threshold を
  wr_byte に渡す箇所を byte にキャスト。

+++ src/monster/monster-damage.cpp
@@ -425,7 +425,7 @@ void MonsterDamageProcessor::get_exp_from_mon(const CreatureEntity &target, int 
      * - Varying speed effects
      * - Get a fraction in proportion of damage point
      */
-    auto new_exp = monrace.level * speed_to_energy(target.speed) * exp_dam;
+    auto new_exp = monrace.level * speed_to_energy(static_cast<byte>(target.speed)) * exp_dam;
     auto new_exp_frac = 0U;
     auto div_h = 0;
     auto div_l = (uint)((creature.max_plv + 2) * speed_to_energy(monrace.speed));
